image-requests-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.10.1
beanie==1.26.0
pytesseract==0.3.13
pillow==10.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
prometheus-fastapi-instrumentator==6.1.0
python-dotenv==1.0.1
pytest==8.3.3
pytest-cov==5.0.0
httpx==0.27.2
pydantic-settings==2.5.2
